+++ .github/workflows/main.yml
@@ -78,4 +78,4 @@ jobs:
           python manage.py migrate --noinput
 
       - name: Run test suite
-        run: python -m pytest llm/tests/ pillars/tests/ -v
+        run: python -m pytest llm/tests/ pillars/tests/ accounts/tests.py -v
