+++ backend/requirements.txt
@@ -1,6 +1,6 @@
 asgiref==3.10.0
 black==25.1.0
-cryptography==44.1.1
+cryptography==44.0.3
 certifi==2025.10.5
 cfgv==3.4.0
 charset-normalizer==3.4.1
